+++ spec/classes/init_spec.rb
@@ -2,6 +2,7 @@ require 'spec_helper'
 
 describe 'ejabberd' do
     it { should contain_class('ejabberd::package') }
+    it { should contain_class('ejabberd::mysql') }
     it { should contain_class('ejabberd::config') }
     it { should contain_class('ejabberd::service') }
 end